+++ apps/server/src/schema/accounts/resolvers/Query/account.ts
@@ -3,7 +3,9 @@ import { db } from '../../../../database';
 import * as authUtils from '../../../../utils/authUtils';
 
 export const account: NonNullable<QueryResolvers['account']> = async (_parent, _arg, { token, user }) => {
-  authUtils.authenticate(token);
+  if (!authUtils.authenticate(token)) {
+    return null;
+  }
 
   const userRecord = await db.accounts.findByPk(user.accountId, { plain: true });
   if (!userRecord || userRecord.dataValues.accountType !== 'superuser') {