+++ apps/server/types/server.d.ts
@@ -1,4 +1,10 @@
 export interface RequestContext {
+  req: any;
   res: any;
   token: string;
+  user: {
+    accountId?: number;
+    email?: string;
+    [key: string]: any;
+  };
 }